[Fix] fix dataclass serialization issue (make ParsedMenu.to_dict() )
